+++ dirty_cat/super_vectorizer.py
@@ -394,7 +394,10 @@ class SuperVectorizer(ColumnTransformer):
         """
         if _sklearn_loose_version < LooseVersion('0.23'):
             try:
-                ct_feature_names = super().get_feature_names()
+                if _sklearn_loose_version < LooseVersion('1.0'):
+                    ct_feature_names = super().get_feature_names()
+                else:
+                    ct_feature_names = super().get_feature_names_out()
             except NotImplementedError:
                 raise NotImplementedError(
                     'Prior to sklearn 0.23, get_feature_names with '
@@ -403,7 +406,10 @@ class SuperVectorizer(ColumnTransformer):
                     'transformers, or update your copy of scikit-learn.'
                 )
         else:
-            ct_feature_names = super().get_feature_names()
+            if _sklearn_loose_version < LooseVersion('1.0'):
+                ct_feature_names = super().get_feature_names()
+            else:
+                ct_feature_names = super().get_feature_names_out()
         all_trans_feature_names = []
 
         for name, trans, cols, _ in self._iter(fitted=True):
@@ -426,3 +432,10 @@ class SuperVectorizer(ColumnTransformer):
             return ct_feature_names
 
         return all_trans_feature_names
+    
+    def get_feature_names_out(self, input_features=None) -> List[str]:
+        """
+        Ensures compatibility with sklearn >= 1.0, and returns the output of
+        get_feature_names.
+        """
+        return self.get_feature_names()